# src/main/resources/swagger/swagger-api.json
{
    "swagger": "2.0",
    "info": {
      "title": "Trouble Ticket",
      "description": "This is Swagger UI environment generated for the TMF Trouble Ticket specification",
      "version": "2.0"
    },
    "host": "serverRoot",
    "basePath": "/tmf-api/troubleTicket/v2/",
    "schemes": [
      "https"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
      "/troubleTicket": {
        "get": {
          "operationId": "listTroubleTicket",
          "summary": "List or find 'TroubleTicket' objects",
          "tags": [
            "troubleTicket"
          ],
          "parameters": [
            {
              "type": "string",
              "required": false,
              "in": "query",
              "name": "fields",
              "description": "Comma separated properties to display in response"
            },
            {
              "type": "integer",
              "required": false,
              "in": "query",
              "name": "offset",
              "description": "Requested index for start of resources to be provided in response"
            },
            {
              "type": "integer",
              "required": false,
              "in": "query",
              "name": "limit",
              "description": "Requested number of resources to be provided in response"
            }
          ],
          "responses": {
            "200": {
              "description": "Ok",
              "schema": {
                "items": {
                  "$ref": "#/definitions/TroubleTicket"
                },
                "type": "array"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "401": {
              "description": "Unauthorized",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "403": {
              "description": "Forbidden",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "405": {
              "description": "Method Not allowed",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "409": {
              "description": "Conflict",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        },
        "post": {
          "operationId": "createTroubleTicket",
          "summary": "Creates a 'TroubleTicket'",
          "tags": [
            "troubleTicket"
          ],
          "parameters": [
            {
              "schema": {
                "$ref": "#/definitions/TroubleTicket_Create"
              },
              "required": true,
              "in": "body",
              "name": "troubleTicket",
              "description": "The Trouble Ticket to be created"
            }
          ],
          "responses": {
            "201": {
              "description": "Created",
              "schema": {
                "$ref": "#/definitions/TroubleTicket"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "401": {
              "description": "Unauthorized",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "403": {
              "description": "Forbidden",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "405": {
              "description": "Method Not allowed",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "409": {
              "description": "Conflict",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/troubleTicket/{id}": {
        "get": {
          "operationId": "retrieveTroubleTicket",
          "summary": "Retrieves a 'TroubleTicket' by Id",
          "tags": [
            "troubleTicket"
          ],
          "parameters": [
            {
              "required": true,
              "type": "string",
              "name": "id",
              "in": "path",
              "description": "Identifier of the Trouble Ticket"
            }
          ],
          "responses": {
            "200": {
              "description": "Ok",
              "schema": {
                "items": {
                  "$ref": "#/definitions/TroubleTicket"
                },
                "type": "array"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "401": {
              "description": "Unauthorized",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "403": {
              "description": "Forbidden",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "405": {
              "description": "Method Not allowed",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "409": {
              "description": "Conflict",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        },
        "patch": {
          "operationId": "patchTroubleTicket",
          "summary": "Updates partially a 'TroubleTicket' by Id",
          "tags": [
            "troubleTicket"
          ],
          "parameters": [
            {
              "required": true,
              "type": "string",
              "name": "id",
              "in": "path",
              "description": "Identifier of the Trouble Ticket"
            },
            {
              "schema": {
                "$ref": "#/definitions/TroubleTicket_Update"
              },
              "required": true,
              "in": "body",
              "name": "troubleTicket",
              "description": "The Trouble Ticket to be updated"
            }
          ],
          "responses": {
            "200": {
              "description": "Updated",
              "schema": {
                "$ref": "#/definitions/TroubleTicket"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "401": {
              "description": "Unauthorized",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "403": {
              "description": "Forbidden",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "405": {
              "description": "Method Not allowed",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "409": {
              "description": "Conflict",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        },
        "delete": {
          "operationId": "deleteTroubleTicket",
          "summary": "Deletes a 'TroubleTicket' by Id",
          "tags": [
            "troubleTicket"
          ],
          "parameters": [
            {
              "required": true,
              "type": "string",
              "name": "id",
              "in": "path",
              "description": "Identifier of the Trouble Ticket"
            }
          ],
          "responses": {
            "204": {
              "description": "Deleted"
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "401": {
              "description": "Unauthorized",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "403": {
              "description": "Forbidden",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "405": {
              "description": "Method Not allowed",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "409": {
              "description": "Conflict",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
...
    },
    "definitions": {
      "TroubleTicket": {
        "type": "object",
        "description": "A trouble ticket is a record of an issue that is created, tracked, and managed by a trouble ticket management system",
        "required": [
          "description",
          "severity",
          "ticketType"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the trouble ticket"
          },
          "href": {
            "type": "string",
            "description": "Hyperlink, a reference to the trouble ticket entity"
          },
          "name": {
            "type": "string",
            "description": "Name of the trouble ticket, typically a short description provided by the user that create the ticket"
          },
          "description": {
            "type": "string",
            "description": "Description of the trouble or issue"
          },
          "externalId": {
            "type": "string",
            "description": "Additional identifier coming from an external system"
          },
          "ticketType": {
            "type": "string",
            "description": "represent a business type of the trouble ticket e.g. incident, complain, request"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time",
            "description": "The date on which the trouble ticket was created"
          },
          "lastUpdate": {
            "type": "string",
            "format": "date-time",
            "description": "The date and time that the trouble ticked was last updated"
          },
          "severity": {
            "type": "string",
            "description": "The severity of the issue. Indicate the implication of the issue on the expected functionality e.g. of a system, application, service etc.. \nSeverity values can be for example : Critical, Major, Minor"
          },
          "priority": {
            "type": "string",
            "description": "The priority of the trouble ticket and how quickly the issue should be resolved. Example: Critical, High, Medium, Low. The value is set by the ticket management system considering the severity, ticket type etc..."
          },
          "requestedResolutionDate": {
            "type": "string",
            "format": "date-time",
            "description": "The resolution date requested by the user"
          },
          "expectedResolutionDate": {
            "type": "string",
            "format": "date-time",
            "description": "The expected resolution date determined by the trouble ticket system"
          },
          "resolutionDate": {
            "type": "string",
            "format": "date-time",
            "description": "The date and time the trouble ticket was resolved"
          },
          "status": {
            "type": "string",
            "description": "The current status of the trouble ticket"
          },
          "statusChangeReason": {
            "type": "string",
            "description": "The reason for changing the status"
          },
          "@type": {
            "type": "string",
            "description": "The (class) type of the trouble ticket. e.g. billingTicket, networkTicket, resourceTicket"
          },
          "@baseType": {
            "type": "string",
            "description": "The base type (class) of the resource. Here can be 'TroubleTicket'"
          },
          "@schemaLocation": {
            "type": "string",
            "description": "Link to the schema describing this REST resource"
          },
          "relatedEntity": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/RelatedEntityRef"
            },
            "description": "An entity that is related to the ticket such as a bill, a product, etc. The entity against which the ticket is associated."
          },
          "channel": {
            "$ref": "#/definitions/Channel",
            "description": "The channel that origin the trouble ticket"
          },
          "statusChange": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/StatusChange"
            },
            "description": "The status change history that are associated to the ticket.Populated by the server"
          },
          "note": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Note"
            },
            "description": "The note(s) that are associated to the ticket."
          },
          "relatedParty": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/RelatedPartyRef"
            },
            "description": "The related party(ies) that are associated to the ticket."
          },
          "ticketRelationship": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/TicketRelationship"
            },
            "description": "A list of trouble ticket relationships (TroubleTicketRelationship [*]). Represents a relationship between trouble tickets"
          },
          "attachment": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Attachment"
            },
            "description": "File(s) attached to the trouble ticket. e.g. pictur of broken device, scaning of a bill or charge"
          }
        }
      },
      "TroubleTicket_Create": {
        "type": "object",
        "description": "A trouble ticket is a record of an issue that is created, tracked, and managed by a trouble ticket management system\nSkipped properties: id,href",
        "required": [
          "description",
          "severity",
          "ticketType"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the trouble ticket, typically a short description provided by the user that create the ticket"
          },
          "description": {
            "type": "string",
            "description": "Description of the trouble or issue"
          },
          "externalId": {
            "type": "string",
            "description": "Additional identifier coming from an external system"
          },
          "ticketType": {
            "type": "string",
            "description": "represent a business type of the trouble ticket e.g. incident, complain, request"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time",
            "description": "The date on which the trouble ticket was created"
          },
          "lastUpdate": {
            "type": "string",
            "format": "date-time",
            "description": "The date and time that the trouble ticked was last updated"
          },
          "severity": {
            "type": "string",
            "description": "The severity of the issue. Indicate the implication of the issue on the expected functionality e.g. of a system, application, service etc.. \nSeverity values can be for example : Critical, Major, Minor"
          },
          "priority": {
            "type": "string",
            "description": "The priority of the trouble ticket and how quickly the issue should be resolved. Example: Critical, High, Medium, Low. The value is set by the ticket management system considering the severity, ticket type etc..."
          },
          "requestedResolutionDate": {
            "type": "string",
            "format": "date-time",
            "description": "The resolution date requested by the user"
          },
          "expectedResolutionDate": {
            "type": "string",
            "format": "date-time",
            "description": "The expected resolution date determined by the trouble ticket system"
          },
          "resolutionDate": {
            "type": "string",
            "format": "date-time",
            "description": "The date and time the trouble ticket was resolved"
          },
          "status": {
            "type": "string",
            "description": "The current status of the trouble ticket"
          },
          "statusChangeReason": {
            "type": "string",
            "description": "The reason for changing the status"
          },
          "@type": {
            "type": "string",
            "description": "The (class) type of the trouble ticket. e.g. billingTicket, networkTicket, resourceTicket"
          },
          "@baseType": {
            "type": "string",
            "description": "The base type (class) of the resource. Here can be 'TroubleTicket'"
          },
          "@schemaLocation": {
            "type": "string",
            "description": "Link to the schema describing this REST resource"
          },
          "relatedEntity": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/RelatedEntityRef"
            },
            "description": "An entity that is related to the ticket such as a bill, a product, etc. The entity against which the ticket is associated."
          },
          "channel": {
            "$ref": "#/definitions/Channel",
            "description": "The channel that origin the trouble ticket"
          },
          "statusChange": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/StatusChange"
            },
            "description": "The status change history that are associated to the ticket.Populated by the server"
          },
          "note": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Note"
            },
            "description": "The note(s) that are associated to the ticket."
          },
          "relatedParty": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/RelatedPartyRef"
            },
            "description": "The related party(ies) that are associated to the ticket."
          },
          "ticketRelationship": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/TicketRelationship"
            },
            "description": "A list of trouble ticket relationships (TroubleTicketRelationship [*]). Represents a relationship between trouble tickets"
          },
          "attachment": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Attachment"
            },
            "description": "File(s) attached to the trouble ticket. e.g. picture of broken device, scanning of a bill or charge"
          }
        }
      },
      "TroubleTicket_Update": {
        "type": "object",
        "description": "A trouble ticket is a record of an issue that is created, tracked, and managed by a trouble ticket management system\nSkipped properties: id,href,id,href,creationDate,lastUpdated,statusChange,@baseType,@type,@schemaLocation",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the trouble ticket, typically a short description provided by the user that create the ticket"
          },
          "description": {
            "type": "string",
            "description": "Description of the trouble or issue"
          },
          "externalId": {
            "type": "string",
            "description": "Additional identifier coming from an external system"
          },
          "ticketType": {
            "type": "string",
            "description": "represent a business type of the trouble ticket e.g. incident, complain, request"
          },
          "severity": {
            "type": "string",
            "description": "The severity of the issue. Indicate the implication of the issue on the expected functionality e.g. of a system, application, service etc.. \nSeverity values can be for example : Critical, Major, Minor"
          },
          "priority": {
            "type": "string",
            "description": "The priority of the trouble ticket and how quickly the issue should be resolved. Example: Critical, High, Medium, Low. The value is set by the ticket management system considering the severity, ticket type etc..."
          },
          "requestedResolutionDate": {
            "type": "string",
            "format": "date-time",
            "description": "The resolution date requested by the user"
          },
          "expectedResolutionDate": {
            "type": "string",
            "format": "date-time",
            "description": "The expected resolution date determined by the trouble ticket system"
          },
          "resolutionDate": {
            "type": "string",
            "format": "date-time",
            "description": "The date and time the trouble ticket was resolved"
          },
          "status": {
            "type": "string",
            "description": "The current status of the trouble ticket"
          },
          "statusChangeReason": {
            "type": "string",
            "description": "The reason for changing the status"
          },
          "relatedEntity": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/RelatedEntityRef"
            },
            "description": "An entity that is related to the ticket such as a bill, a product, etc. The entity against which the ticket is associated."
          },
          "channel": {
            "$ref": "#/definitions/Channel",
            "description": "The channel that origin the trouble ticket"
          },
          "note": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Note"
            },
            "description": "The note(s) that are associated to the ticket."
          },
          "relatedParty": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/RelatedPartyRef"
            },
            "description": "The related party(ies) that are associated to the ticket."
          },
          "ticketRelationship": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/TicketRelationship"
            },
            "description": "A list of trouble ticket relationships (TroubleTicketRelationship [*]). Represents a relationship between trouble tickets"
          },
          "attachment": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Attachment"
            },
            "description": "File(s) attached to the trouble ticket. e.g. pictur of broken device, scaning of a bill or charge"
          },
          "@type": {
            "type": "string",
            "description": "The class type of the actual resource (for type extension)."
          },
          "@schemaLocation": {
            "type": "string",
            "description": "A link to the schema describing a resource (for type extension)."
          },
          "@baseType": {
            "type": "string",
            "description": "The base type for use in polymorphic collections"
          }
        }
      },
...
      "Channel": {
        "type": "object",
        "description": "The communication channel for example WEB, Call Center",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the channel"
          },
          "name": {
            "type": "string",
            "description": "The name of the channel"
          },
          "@type": {
            "type": "string",
            "description": "Indicates the (class) type of channel"
          }
        }
      },
...
      "Error": {
        "required": [
          "code",
          "reason"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "description": "Application related code."
          },
          "reason": {
            "type": "integer",
            "description": "Text that explains the reason for error."
          },
          "message": {
            "type": "string",
            "description": "(optional) Text that provide more details and corrective actions related to the error."
          },
          "status": {
            "type": "integer",
            "description": "(optional) http error code extension like 400-2"
          },
          "referenceError": {
            "type": "string",
            "description": "(optional) A URL to online documentation that provides more information about the error."
          },
          "@type": {
            "type": "string",
            "description": "(optional) The class type of a REST resource."
          },
          "@schemaLocation": {
            "type": "string",
            "description": "(optional) A link to the schema describing a REST resource."
          }
        }
      }
    }
  }